% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averages.R
\name{averages}
\alias{averages}
\title{Averages}
\usage{
averages(df, n_MA, n_quick_MACD, n_slow_MACD, n_signal_MACD)
}
\arguments{
\item{df}{dataframe}

\item{n_MA}{Window Moving Average}

\item{n_quick_MACD}{quick MACD}

\item{n_slow_MACD}{slow MACD}

\item{n_signal_MACD}{signal MACD}
}
\value{
dataframe with the financial indicators added
}
\description{
Function to  add financial indicators into the dataset
}
\details{
This function has been designed to add financial indicators into the dataset (Moving average and MACD)
}
