% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastPriceMultiplePlatforms.R
\name{getLastPriceMultiplePlatforms}
\alias{getLastPriceMultiplePlatforms}
\title{getLastPriceMultiplePlatforms}
\usage{
getLastPriceMultiplePlatforms(cryptocurrency)
}
\arguments{
\item{cryptocurrency}{cryptocurrency id}
}
\value{
dataframe with prices on six platforms
}
\description{
a function used to get the prices of the imputted crypto on 6 platforms
}
