% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastweek_minute.R
\name{lastweek_minute}
\alias{lastweek_minute}
\title{lastweek_minute}
\usage{
lastweek_minute(crytocurrenty = "BTC", comparison = "USD")
}
\arguments{
\item{crytocurrenty}{cryptocurrency to analyse}

\item{comparison}{currency to be compared}
}
\value{
dataframe with the time, highest price, lowest price, open price, close price of the chosen timeframe

dataframe with all the information required
}
\description{
Function to get the information from the cryptocompare API per minute
}
\details{
This function has been designed to get a dataframe from the crytocompare API with the highest, lowest, open and close price from the crytocurrency chosen in minute of the last week.
}
